+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./register-worker";
 export * from "./elastic-worker";
 export * from "./errors";
 export { Transfer } from "./utils/transfer";
+export { TaskStore, Task, TaskStoreConstructor } from "./types";